requirements.txt
tk
psycopg2-binary
requests
pillow
pywebview
python-dotenv
bcrypt
tkinterweb
cefpython3

